Backend/requirements.txt
qrcode
django-cors-headers
